requirements.txt
yfinance==0.2.54
pandas==2.2.3
pandas-ta==0.3.14b
numpy==1.26.4
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
python-dotenv==1.1.0
chronos-forecasting==1.5.0
torch==2.6.0
requests==2.32.3
apscheduler==3.11.0
